refactor: 파일명 변경에 따른 경로 변경 #9
