+++ src/pages/RepaymentSimulationResult.vue
@@ -91,7 +91,7 @@
 <script setup>
 import { onMounted } from 'vue';
 import { useRoute, useRouter } from 'vue-router';
-import { useSimulationStore } from '@/stores/simulation';
+import { useSimulationStore } from '@/stores/repayment-simulation';
 import DebtChart from '@/components/domain/simulation/DebtChart.vue';
 import PlanModal from '@/components/domain/simulation/PlanModal.vue';
 import styles from '@/assets/styles/pages/simulation/RepaymentSimulationResult.module.css';
