+++ src/pages/Simulation.vue
@@ -1,5 +1,5 @@
 <script setup>
-import SimulationTabMenu from '@/components/domain/Simulation/SimulationTabMenu.vue';
+import SimulationTabMenu from '@/components/domain/simulation/SimulationTabMenu.vue';
 </script>
 
 <template>
